Reverse a Linked List
Given a linked list, the task is to reverse the linked list by changing the links between nodes.

Examples: 

Input: head: 1 -> 2 -> 3 -> 4 -> NULL
Output: head: 4 -> 3 -> 2 -> 1 -> NULL
Explanation: Reversed Linked List:

Reverse-a-Linked-List-2


Input: head: 1 -> 2 -> 3 -> 4 -> 5 -> NULL
Output: head: 5 -> 4 -> 3 -> 2 -> 1 -> NULL
Explanation: Reversed Linked List:

Using Stack - O(n) Time and O(n) Space
The idea is to traverse the linked list and push all nodes except the last node into the stack. Make the last node as the new head of the reversed linked list. Now, start popping the element and append each node to the reversed Linked List. Finally, return the head of the reversed linked list.


Follow the steps below to solve the problem:

Push all the nodes(values and address) except the last node in the stack.
Once the nodes are pushed, update the Head pointer to the last node.
Start popping the nodes and push them at the end of the linked list in the same order until the stack is empty.
Update the next pointer of last node in the stack by NULL.